Port the ledger sorting example from the old samples solution
Most of the code was copied verbatim from DotNetCoreSamples. I put it in with
the existing ListSorting code, but most of it is in the WPF app because it
uses the CollectionView classes from System.Windows.Data, which I want to keep
out of the base project.

I did not create a console version of the sample because I am still not clear
on how to sort an ObservableCollection in that environment. If you are meant
to just cast/convert it to a List then I don't see the point of having a
sample.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>