Add example using OnPropertyChanged()
OnPropertyChanged() works as expected. The problem I have had in the past
with List<> is that OnPropertyChanged() checks for a difference somehow
(maybe the listeners do the check), and the List<> "pointer" is not actually
changing, so no update to the UI occurs. I have not found a way to "force"
the update; the only workaround is to "toggle" the property to null and back.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>